backend/requirements.txt
requests
pandas
python-dotenv
notebook
matplotlib
scikit-learn
